feat: add is_seed column and refactor the template table.

+++ src/main/scala/jp/ijufumi/openreports/services/impl/ReportServiceImpl.scala
@@ -162,7 +162,9 @@ class ReportServiceImpl @Inject() (
     }
     val template = templateOpt.get
     templateRepository.delete(db, workspaceId, id)
-    storageService.delete(workspaceId, template.filePath, template.storageType)
+    if (!template.isSeed) {
+      storageService.delete(workspaceId, template.filePath, template.storageType)
+    }
   }
 
   override def getGroups(workspaceId: String, page: Int, limit: Int): Lists[ReportGroup] = {
